Work around Get DCMI Capabilities Info spec violation
Some BMCs (SuperMicro) claim a v1.1 Mandatory Platform Attributes
response, however return a v1.0-formatted body. We work around this
by assuming v1.0 if the body is 4 bytes as a special case,
otherwise believing what it says. This could be improved by also
looking at whether bytes 3 and 4, which are reserved in v1.1 and
v1.5, are null.