Allow configuring privilege level when creating a version-agnostic session
More generally, this allows arbitrary options common to v1.5 and
v2.0 to be specified using the agnostic API. Previously this could
only be set when the IPMI version was known, which should not be the
case. The primary goal is agnosticity, not simplicity - if users
do not want to specify a privilege level, they can omit the option.