Improve implementation of Get Channel Cipher Suites
Notably, this changes V2SessionOpts to accept a slice of suites rather than a
trio of algorithms. The zero value must be useful to most people, which means
we cannot default to cipher suite 0, so we have no choice but to make this
struct a more high-level API. Fortunately, people do not generally want to mess
around with session establishment - they want the flexibility in commands they
send. Also for this reason, we do not expose the underlying command, it is only
available via the RetrieveSupportedCipherSuites() wrapper, which is the only
practical use.

Closes #50.